+++ lib/sbmt/kafka_consumer/testing/shared_contexts/with_sbmt_karafka_consumer.rb
@@ -22,6 +22,7 @@ RSpec.shared_context "with sbmt karafka consumer" do
   }
 
   before {
+    Sbmt::KafkaConsumer::ClientConfigurer.configure!
     allow(kafka_client).to receive(:mark_as_consumed!).and_return(true)
   }