+++ spec/sbmt/kafka_consumer/instrumentation/yabeda_metrics_listener_spec.rb
@@ -6,6 +6,10 @@ describe Sbmt::KafkaConsumer::Instrumentation::YabedaMetricsListener do
   let(:message) { build(:messages_message) }
   let(:messages) { OpenStruct.new(metadata: build(:messages_batch_metadata), count: 1) }
 
+  before do
+    Sbmt::KafkaConsumer::ClientConfigurer.configure!
+  end
+
   describe ".on_statistics_emitted" do
     let(:base_rdkafka_stats) {
       {